feat: add basic auth middleware

+++ cmd/server/accounts.go
@@ -5,6 +5,7 @@ import (
 	accountsv1 "buf.build/gen/go/xskydev/go-money-pb/protocolbuffers/go/gomoneypb/accounts/v1"
 	"connectrpc.com/connect"
 	"context"
+	"github.com/ft-t/go-money/pkg/auth"
 	"github.com/ft-t/go-money/pkg/boilerplate"
 )
 
@@ -16,7 +17,10 @@ func (a *AccountsApi) CreateAccountsBulk(
 	ctx context.Context,
 	c *connect.Request[accountsv1.CreateAccountsBulkRequest],
 ) (*connect.Response[accountsv1.CreateAccountsBulkResponse], error) {
-	// todo auth
+	jwtData := auth.FromContext(ctx)
+	if jwtData.UserID == 0 {
+		return nil, connect.NewError(connect.CodeUnauthenticated, auth.ErrInvalidToken)
+	}
 
 	resp, err := a.accSvc.CreateBulk(ctx, c.Msg)
 	if err != nil {
@@ -29,13 +33,19 @@ func (a *AccountsApi) CreateAccountsBulk(
 }
 
 func (a *AccountsApi) ReorderAccounts(ctx context.Context, c *connect.Request[accountsv1.ReorderAccountsRequest]) (*connect.Response[accountsv1.ReorderAccountsResponse], error) {
-
+	jwtData := auth.FromContext(ctx)
+	if jwtData.UserID == 0 {
+		return nil, connect.NewError(connect.CodeUnauthenticated, auth.ErrInvalidToken)
+	}
 	//TODO implement me
 	panic("implement me")
 }
 
 func (a *AccountsApi) DeleteAccount(ctx context.Context, c *connect.Request[accountsv1.DeleteAccountRequest]) (*connect.Response[accountsv1.DeleteAccountResponse], error) {
-	// todo auth
+	jwtData := auth.FromContext(ctx)
+	if jwtData.UserID == 0 {
+		return nil, connect.NewError(connect.CodeUnauthenticated, auth.ErrInvalidToken)
+	}
 
 	resp, err := a.accSvc.Delete(ctx, c.Msg)
 	if err != nil {
@@ -49,7 +59,10 @@ func (a *AccountsApi) CreateAccount(
 	ctx context.Context,
 	c *connect.Request[accountsv1.CreateAccountRequest],
 ) (*connect.Response[accountsv1.CreateAccountResponse], error) {
-	// todo auth
+	jwtData := auth.FromContext(ctx)
+	if jwtData.UserID == 0 {
+		return nil, connect.NewError(connect.CodeUnauthenticated, auth.ErrInvalidToken)
+	}
 
 	resp, err := a.accSvc.Create(ctx, c.Msg)
 	if err != nil {
@@ -63,7 +76,10 @@ func (a *AccountsApi) UpdateAccount(
 	ctx context.Context,
 	c *connect.Request[accountsv1.UpdateAccountRequest],
 ) (*connect.Response[accountsv1.UpdateAccountResponse], error) {
-	// todo auth
+	jwtData := auth.FromContext(ctx)
+	if jwtData.UserID == 0 {
+		return nil, connect.NewError(connect.CodeUnauthenticated, auth.ErrInvalidToken)
+	}
 
 	resp, err := a.accSvc.Update(ctx, c.Msg)
 	if err != nil {
@@ -77,7 +93,11 @@ func (a *AccountsApi) ListAccounts(
 	ctx context.Context,
 	c *connect.Request[accountsv1.ListAccountsRequest],
 ) (*connect.Response[accountsv1.ListAccountsResponse], error) {
-	// todo auth
+	jwtData := auth.FromContext(ctx)
+	if jwtData.UserID == 0 {
+		return nil, connect.NewError(connect.CodeUnauthenticated, auth.ErrInvalidToken)
+	}
+
 	resp, err := a.accSvc.List(ctx, c.Msg)
 	if err != nil {
 		return nil, err
